# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight backend RESTful API designed for tracking and managing personal or organizational expenses. This API provides secure endpoints to log daily expenses, categorize spending, and generate basic financial summaries.",
        "title": "API Backend Expense",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/expenses": {
            "get": {
                "description": "get all existing expense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "List expense",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "post": {
                "description": "create expense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Add new expense",
                "parameters": [
                    {
                        "description": "Create new expense payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateExpensePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            }
        },
        "/expenses/{id}": {
            "get": {
                "description": "get one filters by expense id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Get an expense",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Income ID (Optional)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "put": {
                "description": "update expense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Edit expense",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Income ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Edit expense payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateExpensePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete expense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Remove expense",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Income ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateExpensePayload": {
            "type": "object",
            "required": [
                "amount",
                "category_id",
                "expense_date",
                "title"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "expense_date": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "note": {
                    "type": "string",
                    "maxLength": 255
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "dto.Meta": {
            "type": "object",
            "properties": {
                "paginate": {
                    "$ref": "#/definitions/dto.Paginate"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "dto.Paginate": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "dto.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/apperror.AppError"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/dto.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateExpensePayload": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "expense_date": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "note": {
                    "type": "string",
                    "maxLength": 255
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
    }
}